Modifies samples to be self contained rather than calling print function from runReport
