+++ samples/runReportWithMultipleMetrics.js
@@ -69,13 +69,36 @@ function main(propertyId = 'YOUR-GA4-PROPERTY-ID') {
         },
       ],
     });
-
-    // Imports the runReport sample used for printing results
-    const runReport = require('./runReport.js');
-    runReport.printRunReportResponse(response);
+    printRunReportResponse(response);
   }
 
   runReportWithMultipleMetrics();
+
+  // Prints results of a runReport call.
+  function printRunReportResponse(response) {
+    //[START analyticsdata_print_run_report_response_header]
+    console.log(response.rowCount + ' rows received');
+    response.dimensionHeaders.forEach(dimensionHeader => {
+      console.log('Dimension header name: ' + dimensionHeader.name);
+    });
+    response.metricHeaders.forEach(metricHeader => {
+      console.log(
+        'Metric header name: ' +
+          metricHeader.name +
+          ' (' +
+          metricHeader.type +
+          ')'
+      );
+    });
+    //[END analyticsdata_print_run_report_response_header]
+
+    // [START analyticsdata_print_run_report_response_rows]
+    console.log('Report result:');
+    response.rows.forEach(row => {
+      console.log(row.dimensionValues[0].value, ',', row.metricValues[0].value);
+    });
+    // [END analyticsdata_print_run_report_response_rows]
+  }
   // [END analyticsdata_run_report_with_multiple_metrics]
 }
